refactor: separate the visualization code to serve as template for next algorithms
