+++ src/components/bubble-sort/tutorial.tsx
@@ -7,7 +7,12 @@ import {
 	DialogTrigger
 } from "@/components/ui/dialog";
 
-const Tutorial: React.FC = () => {
+interface TutorialProps {
+	title: string;
+	description: string;
+	steps: string[];
+}
+const Tutorial: React.FC<TutorialProps> = ({ title, description, steps }) => {
 	return (
 		<Dialog>
 			<DialogTrigger asChild>
@@ -17,40 +22,18 @@ const Tutorial: React.FC = () => {
 			</DialogTrigger>
 			<DialogContent className="bg-white text-gray-800">
 				<DialogHeader>
-					<DialogTitle className="text-2xl font-bold">
-						Bubble Sort Tutorial
-					</DialogTitle>
+					<DialogTitle className="text-2xl font-bold">{title}</DialogTitle>
 					<DialogDescription className="text-gray-600">
-						Bubble Sort is a simple sorting algorithm that repeatedly steps
-						through the list, compares adjacent elements and swaps them if they
-						are in the wrong order.
+						{description}
 					</DialogDescription>
 				</DialogHeader>
 				<div className="mt-4">
 					<h3 className="text-lg font-semibold mb-2">Algorithm Steps:</h3>
 					<ol className="list-decimal list-inside space-y-2">
-						<li>Start with an unsorted array of n elements.</li>
-						<li>
-							Compare adjacent elements, swapping them if they are in the wrong
-							order.
-						</li>
-						<li>
-							Repeat step 2 for each pair of adjacent elements, from the
-							beginning of the array to the end.
-						</li>
-						<li>
-							After each pass, the largest unsorted element &quot;bubbles
-							up&quot; to its correct position.
-						</li>
-						<li>
-							Repeat steps 2-4 for n-1 passes, where n is the number of elements
-							in the array.
-						</li>
+						{steps.map((step, index) => (
+							<li key={index}>{step}</li>
+						))}
 					</ol>
-					<p className="mt-4">
-						The algorithm gets its name from the way smaller elements
-						&quot;bubble&quot; to the top of the list with each iteration.
-					</p>
 				</div>
 			</DialogContent>
 		</Dialog>
